devel/git: Drop git-gitk from git metapakage
As discussed at length on pkgsrc-users: A number of people believe
that the git metapackage should contain all of what is in git's
release tarball, and that people should read DESCR.  A number of
people believe that the package named git should contain less (not
gitk), and sometimes even less than that.  As a compromise where each
group is merely grumbling a little, drop gitk from the metapackage,
because it is a component that some are bothered by (because it
depends on tk, tcl and X, mostly), and because it is easy for those
that want it to install it separately.