gdk-pixbuf2: Couple of SunOS fixes.
Using _XOPEN_SOURCE=500 is insufficient when compiling in C99 mode that this
package specifies.  600 is the minimum required for C99, so bump to that.

Meson expects the files in build-aux to be executable, but they aren't as
shipped in the source.  I don't understand how nobody else is seeing this,
but explicitly marking them as executable fixes SunOS and won't hurt others.